Add a body field to requests
Adds a migration to add a body field to freeagent_api_requests table
and starts reshaping the code to allow users to make requests that
include a body. Make improvements to the trigger.html.erb view and
requests helper to be able to deal with a scenario where the response
doesn't include body.
Update README to remove a line saying only GET requests are handled.